# frontend/src/api/openapi.json
{
  "openapi": "3.1.0",
  "info": {
    "title": "Futbol Predict API",
    "description": "API local para prediccion probabilistica 1X2.",
    "version": "0.1.0"
  },
  "paths": {
...
    "/backtests/football-data-uk": {
      "get": {
        "summary": "Football Data Uk Backtest",
        "operationId": "football_data_uk_backtest_backtests_football_data_uk_get",
        "parameters": [
          {
            "name": "season",
            "in": "query",
            "required": false,
            "schema": {
              "type": "string",
              "pattern": "^\\d{4}$",
              "default": "2526",
              "title": "Season"
            }
          },
          {
            "name": "division",
            "in": "query",
            "required": false,
            "schema": {
              "type": "string",
              "minLength": 2,
              "maxLength": 4,
              "default": "E0",
              "title": "Division"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/BacktestResponse"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/backtests/football-data-uk/big-five": {
      "get": {
        "summary": "Football Data Uk Big Five Backtest",
        "operationId": "football_data_uk_big_five_backtest_backtests_football_data_uk_big_five_get",
        "parameters": [
          {
            "name": "start_season",
            "in": "query",
            "required": false,
            "schema": {
              "type": "string",
              "pattern": "^\\d{4}$",
              "default": "1617",
              "title": "Start Season"
            }
          },
          {
            "name": "end_season",
            "in": "query",
            "required": false,
            "schema": {
              "type": "string",
              "pattern": "^\\d{4}$",
              "default": "2526",
              "title": "End Season"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/BacktestResponse"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/backtests/db/football-data-uk": {
      "get": {
        "summary": "Db Football Data Uk Backtest",
        "operationId": "db_football_data_uk_backtest_backtests_db_football_data_uk_get",
        "parameters": [
          {
            "name": "season",
            "in": "query",
            "required": false,
            "schema": {
              "type": "string",
              "pattern": "^\\d{4}$",
              "default": "2526",
              "title": "Season"
            }
          },
          {
            "name": "division",
            "in": "query",
            "required": false,
            "schema": {
              "type": "string",
              "minLength": 2,
              "maxLength": 4,
              "default": "E0",
              "title": "Division"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/BacktestResponse"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/backtests/db/big-five": {
      "get": {
        "summary": "Db Big Five Backtest",
        "operationId": "db_big_five_backtest_backtests_db_big_five_get",
        "parameters": [
          {
            "name": "start_season",
            "in": "query",
            "required": false,
            "schema": {
              "type": "string",
              "pattern": "^\\d{4}$",
              "default": "1617",
              "title": "Start Season"
            }
          },
          {
            "name": "end_season",
            "in": "query",
            "required": false,
            "schema": {
              "type": "string",
              "pattern": "^\\d{4}$",
              "default": "2526",
              "title": "End Season"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/BacktestResponse"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
...
    "/calibration/status": {
      "get": {
        "summary": "Calibration Status",
        "operationId": "calibration_status_calibration_status_get",
        "parameters": [
          {
            "name": "bins",
            "in": "query",
            "required": false,
            "schema": {
              "type": "integer",
              "maximum": 50,
              "minimum": 1,
              "default": 10,
              "title": "Bins"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/CalibrationStatusResponse"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/calibration/curves": {
      "get": {
        "summary": "Calibration Curves",
        "operationId": "calibration_curves_calibration_curves_get",
        "parameters": [
          {
            "name": "bins",
            "in": "query",
            "required": false,
            "schema": {
              "type": "integer",
              "maximum": 50,
              "minimum": 1,
              "default": 10,
              "title": "Bins"
            }
          },
          {
            "name": "model",
            "in": "query",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string",
                  "minLength": 1,
                  "maxLength": 120
                },
                {
                  "type": "null"
                }
              ],
              "title": "Model"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/CalibrationCurveResponse"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/models/rankings": {
      "get": {
        "summary": "Model Rankings",
        "operationId": "model_rankings_models_rankings_get",
        "parameters": [
          {
            "name": "min_matches",
            "in": "query",
            "required": false,
            "schema": {
              "type": "integer",
              "minimum": 1,
              "default": 100,
              "title": "Min Matches"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ModelRankingResponse"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/models/champion": {
      "get": {
        "summary": "Champion Model",
        "operationId": "champion_model_models_champion_get",
        "parameters": [
          {
            "name": "min_matches",
            "in": "query",
            "required": false,
            "schema": {
              "type": "integer",
              "minimum": 1,
              "default": 100,
              "title": "Min Matches"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ModelRankingRowResponse"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/fixtures/upcoming": {
      "get": {
        "summary": "Upcoming Fixtures",
        "operationId": "upcoming_fixtures_fixtures_upcoming_get",
        "parameters": [
          {
            "name": "days",
            "in": "query",
            "required": false,
            "schema": {
              "type": "integer",
              "maximum": 120,
              "minimum": 1,
              "default": 14,
              "title": "Days"
            }
          },
          {
            "name": "limit",
            "in": "query",
            "required": false,
            "schema": {
              "type": "integer",
              "maximum": 200,
              "minimum": 1,
              "default": 50,
              "title": "Limit"
            }
          },
          {
            "name": "divisions",
            "in": "query",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string",
                  "minLength": 2
                },
                {
                  "type": "null"
                }
              ],
              "title": "Divisions"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/UpcomingFixturesResponse"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
...
  },
  "components": {
    "schemas": {
...
      "BacktestResponse": {
        "properties": {
          "scope": {
            "type": "string",
            "title": "Scope"
          },
          "source": {
            "type": "string",
            "title": "Source"
          },
          "start_season": {
            "type": "string",
            "title": "Start Season"
          },
          "end_season": {
            "type": "string",
            "title": "End Season"
          },
          "divisions": {
            "items": {
              "type": "string"
            },
            "type": "array",
            "title": "Divisions"
          },
          "n_matches": {
            "type": "integer",
            "title": "N Matches"
          },
          "metrics": {
            "items": {
              "$ref": "#/components/schemas/BacktestMetricResponse"
            },
            "type": "array",
            "title": "Metrics"
          },
          "breakdowns": {
            "items": {
              "$ref": "#/components/schemas/BacktestBreakdownResponse"
            },
            "type": "array",
            "title": "Breakdowns"
          }
        },
        "type": "object",
        "required": [
          "scope",
          "source",
          "start_season",
          "end_season",
          "divisions",
          "n_matches",
          "metrics",
          "breakdowns"
        ],
        "title": "BacktestResponse"
      },
...
      "CalibrationStatusResponse": {
        "properties": {
          "n_bins": {
            "type": "integer",
            "title": "N Bins"
          },
          "rows": {
            "items": {
              "$ref": "#/components/schemas/CalibrationStatusRowResponse"
            },
            "type": "array",
            "title": "Rows"
          }
        },
        "type": "object",
        "required": [
          "n_bins",
          "rows"
        ],
        "title": "CalibrationStatusResponse"
      },
...
      "HTTPValidationError": {
        "properties": {
          "detail": {
            "items": {
              "$ref": "#/components/schemas/ValidationError"
            },
            "type": "array",
            "title": "Detail"
          }
        },
        "type": "object",
        "title": "HTTPValidationError"
      },
...
    }
  }
}